+++ askbot/conf/minimum_reputation.py
@@ -18,7 +18,7 @@ settings.register(
     livesettings.IntegerValue(
         MIN_REP,
         'MIN_REP_TO_VOTE_UP',
-        default=15,
+        default=5,
         description=_('Upvote')
     )
 )
@@ -27,7 +27,7 @@ settings.register(
     livesettings.IntegerValue(
         MIN_REP,
         'MIN_REP_TO_VOTE_DOWN',
-        default=100,
+        default=50,
         description=_('Downvote')
     )
 )
@@ -36,7 +36,7 @@ settings.register(
     livesettings.IntegerValue(
         MIN_REP,
         'MIN_REP_TO_ANSWER_OWN_QUESTION',
-        default=25,
+        default=5,
         description=_('Answer own question immediately')
     )
 )
@@ -45,7 +45,7 @@ settings.register(
     livesettings.IntegerValue(
         MIN_REP,
         'MIN_REP_TO_ACCEPT_OWN_ANSWER',
-        default=50,
+        default=20,
         description=_('Accept own answer')
     )
 )
@@ -54,7 +54,7 @@ settings.register(
     livesettings.IntegerValue(
         MIN_REP,
         'MIN_REP_TO_FLAG_OFFENSIVE',
-        default=15,
+        default=5,
         description=_('Flag offensive')
     )
 )
@@ -63,7 +63,7 @@ settings.register(
     livesettings.IntegerValue(
         MIN_REP,
         'MIN_REP_TO_LEAVE_COMMENTS',
-        default=50,
+        default=10,
         description=_('Leave comments')
     )
 )
@@ -72,7 +72,7 @@ settings.register(
     livesettings.IntegerValue(
         MIN_REP,
         'MIN_REP_TO_DELETE_OTHERS_COMMENTS',
-        default=2000,
+        default=200,
         description=_('Delete comments posted by others')
     )
 )
@@ -81,7 +81,7 @@ settings.register(
     livesettings.IntegerValue(
         MIN_REP,
         'MIN_REP_TO_DELETE_OTHERS_POSTS',
-        default=5000,
+        default=500,
         description=_('Delete questions and answers posted by others')
     )
 )
@@ -90,7 +90,7 @@ settings.register(
     livesettings.IntegerValue(
         MIN_REP,
         'MIN_REP_TO_UPLOAD_FILES',
-        default=60,
+        default=10,
         description=_('Upload files')
     )
 )
@@ -99,7 +99,7 @@ settings.register(
     livesettings.IntegerValue(
         MIN_REP,
         'MIN_REP_TO_CLOSE_OWN_QUESTIONS',
-        default=250,
+        default=25,
         description=_('Close own questions'),
     )
 )
@@ -108,7 +108,7 @@ settings.register(
     livesettings.IntegerValue(
         MIN_REP,
         'MIN_REP_TO_RETAG_OTHERS_QUESTIONS',
-        default=500,
+        default=50,
         description=_('Retag questions posted by other people')
     )
 )
@@ -117,7 +117,7 @@ settings.register(
     livesettings.IntegerValue(
         MIN_REP,
         'MIN_REP_TO_REOPEN_OWN_QUESTIONS',
-        default=500,
+        default=50,
         description=_('Reopen own questions')
     )
 )
@@ -126,7 +126,7 @@ settings.register(
             livesettings.IntegerValue(
         MIN_REP,
         'MIN_REP_TO_EDIT_WIKI',
-        default=750,
+        default=75,
         description=_('Edit community wiki posts')
     )
 )
@@ -135,7 +135,7 @@ settings.register(
     livesettings.IntegerValue(
         MIN_REP,
         'MIN_REP_TO_EDIT_OTHERS_POSTS',
-        default=2000,
+        default=200,
         description=_('Edit posts authored by other people')
     )
 )
@@ -144,7 +144,7 @@ settings.register(
     livesettings.IntegerValue(
         MIN_REP,
         'MIN_REP_TO_VIEW_OFFENSIVE_FLAGS',
-        default=2000,
+        default=200,
         description=_('View offensive flags')
     )
 )
@@ -153,7 +153,7 @@ settings.register(
     livesettings.IntegerValue(
         MIN_REP,
         'MIN_REP_TO_CLOSE_OTHERS_QUESTIONS',
-        default=2000,
+        default=200,
         description=_('Close questions asked by others')
     )
 )
@@ -162,7 +162,7 @@ settings.register(
     livesettings.IntegerValue(
         MIN_REP,
         'MIN_REP_TO_LOCK_POSTS',
-        default=4000,
+        default=400,
         description=_('Lock posts')
     )
 )
@@ -171,7 +171,7 @@ settings.register(
     livesettings.IntegerValue(
         MIN_REP,
         'MIN_REP_TO_HAVE_STRONG_URL',
-        default=250,
+        default=25,
         description=_('Remove rel=nofollow from own homepage'),
         help_text=_(
                     'When a search engine crawler will see a rel=nofollow '