+++ askbot/const/message_keys.py
@@ -33,6 +33,9 @@ _('click to see the most answered questions')
 _('click to see least voted questions')
 _('by votes')
 _('click to see most voted questions')
+_('interesting')
+_('ignored')
+_('subscribed')
 TAGS_ARE_REQUIRED_MESSAGE = _('tags are required')
 TAG_WRONG_CHARS_MESSAGE = _(
     'please use letters, numbers and characters "-+.#"'