+++ askbot/urls.py
@@ -159,6 +159,16 @@ urlpatterns = patterns('',
         views.readers.tags, 
         name='tags'
     ),
+    url(
+        r'^%s$' % _('suggested-tags/'),
+        views.meta.list_suggested_tags,
+        name = 'list_suggested_tags'
+    ),
+    url(#ajax only
+        r'^%s$' % 'moderate-suggested-tag',
+        views.commands.moderate_suggested_tag,
+        name = 'moderate_suggested_tag'
+    ),
     #todo: collapse these three urls and use an extra json data var
     url(#ajax only
         r'^%s%s$' % ('mark-tag/', 'interesting/'),
@@ -261,12 +271,12 @@ urlpatterns = patterns('',
     ),
     url(
         r'^%s$' % _('users/'),
-        views.users.users, 
+        views.users.show_users, 
         name='users'
     ),
     url(
         r'^%s%s(?P<group_id>\d+)/(?P<group_slug>.*)/$' % (_('users/'), _('by-group/')),
-        views.users.users, 
+        views.users.show_users, 
         kwargs = {'by_group': True},
         name = 'users_by_group'
     ),