+++ askbot/doc/source/changelog.rst
@@ -5,16 +5,18 @@ Development version
 -------------------
 * Updated LDAP configuration: allow protocol change, master login and
   adding "extra options" to the ldap session (Evgeny)
-* Simple tag moderation via email (Evgeny)
+* Tag moderation (Evgeny)
 * Editable optional three level category selector for the tags (Evgeny)
 * Tag editor adding tags as they are typed (Evgeny)
 * Welcome email for the case when replying by email is enabled (Evgeny)
-* Detection of email signature based on the response to the welome email (Evgeny)
+* Detection of email signature based on the response to the welcome email (Evgeny)
 * Hide "website" and "about" section of the blocked user profiles
   to help prevent user profile spam (Evgeny)
 * Added a function to create a custom user profile tab,
   the feature requires access to the server (Evgeny)
 * Added optional top banner to the question page (Evgeny)
+* Made "bootstrap mode" default and created instead "large site mode" (Evgeny)
+* Added interesting/ignored/subscribed tags to the user profile page (Paul Backhouse, Evgeny)
 
 0.7.43 (May 14, 2012)
 ---------------------